requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
albucore==0.0.12
albumentations==1.4.11
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
dlib==19.24.4
dnspython==2.6.1
email_validator==2.2.0
eval_type_backport==0.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
gradio==4.31.5
gradio_client==0.16.4
grpcio==1.65.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
imageio==2.34.2
importlib_resources==6.4.0
imutils==0.5.4
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
onnx==1.14.0
onnxruntime==1.15.1
onnxruntime-gpu==1.18.0
onnxsim==0.4.36
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycocotools==2.0.7
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
ruff==0.5.4
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.13.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.3
watchfiles==0.22.0
websockets==11.0.3
Werkzeug==3.0.3
wrapt==1.16.0
playsound==1.3.0
twilio==7.15.0
ultralytics==8.0.42
